カスタム404ページが表示されるように調整した
