+++ src/pages/404.tsx
@@ -1,0 +1,2 @@
+// 404.htmlが生成されないのでpages/404.tsxを設定する
+export { default } from '~/app/not-found';
